Fix view feedback
