+++ src/controllers/profile.controller.ts
@@ -14,6 +14,7 @@ export class ProfileController {
 
     this.getProfile = this.getProfile.bind(this);
     this.editProfile = this.editProfile.bind(this);
+    this.editPicture = this.editPicture.bind(this);
   }
 
   public async getProfile(_: Request, res: Response): Promise<void> {
@@ -57,6 +58,27 @@ export class ProfileController {
       status: "success",
       message: "Profile updated successfully.",
     });
+  }
+
+  public async editPicture(req: Request, res: Response): Promise<void> {
+    const { username } = res.locals.user;
+
+    const user = await this.userRepo.findOne({ where: { username } });
+
+    if (!user) {
+      res.sendStatus(401);
+      return;
+    }
+
+    if (!req.file) {
+      user.picture = null;
+      await this.userRepo.save(user);
+      res.status(200).json({ message: "The picture removed successfully" });
+    } else {
+      user.picture = req.file.buffer;
+      await this.userRepo.save(user);
+      res.status(201).json({ message: "The picture changed successfully" });
+    }
   }
 }
