+++ src/routes/profile.route.ts
@@ -1,3 +1,4 @@
+import multer from "multer";
 import { Router } from "express";
 import { ProfileController } from "../controllers/profile.controller";
 import { DatabaseService } from "../services/database.service";
@@ -7,10 +8,17 @@ export function generateProfileRoutes(
   databaseService: DatabaseService,
 ): Router {
   const router = Router();
+  const upload = multer();
   const controller = new ProfileController(databaseService);
 
   router.get("/", authMiddleware, controller.getProfile);
   router.patch("/", authMiddleware, controller.editProfile);
+  router.post(
+    "/picture",
+    authMiddleware,
+    upload.single("picture"),
+    controller.editPicture,
+  );
 
   return router;
 }
