+++ swagger.json
@@ -327,7 +327,7 @@
         }
       }
     },
-    "/api/picture": {
+    "/api/profile/picture": {
       "post": {
         "summary": "Edit user picture",
         "requestBody": {
